Started adding text to plotting/fitting notebook
I began to add some descriptive text to my simple plotting/fitting
notebook.

First of all, I renamed the notebook to clarify that it is the first one
to be run. I also added many markdown cells of narrative. In addition, I
made a copy of the data file 'kin_data.csv', which has a couple of rows
inverted, so that folks can test another file.

There was some problems with the line spacing for lists, so I modified
custom.css.