Finished marking up the plot/fit notebook
Finally got explanatory text in for the entire first notebook. I made a
change to custom.css to render in-line code a slightly different color.
I also dumped all of the output before saving so that it wouldn't be
uploaded with the notebook.

I was going to add automatic numbering to the MathJax equations, but I
decided against it. Left the commented code in nbprep.py in case I
change my mind later.